+++ test/inspec/author_spec.rb
@@ -46,6 +46,10 @@ describe service('aem-author') do
   it { should_not be_running }
 end
 
+describe file('/opt/aem/author/crx-quickstart/conf/cq.pid') do
+  it { should_not exist }
+end
+
 describe file('/etc/puppetlabs/puppet/author.yaml') do
   it { should be_file }
   it { should exist }